[#67][block_page_not_login] finaly working in router
