+++ front/public/main.js
@@ -10,8 +10,7 @@ import './game/view-game.ce.js';
 import './views/view-reinitialisation-pass-mail.ce.js';
 import './views/view-new-pass.ce.js';
 import './views/view-dashboard.ce.js';
-import './views/authService.js';
-import './views/apiService.js';
+// import './views/authService.js';
 
 
 // router
